add weights

+++ system-parachains/collectives/collectives-polkadot/src/weights/pallet_core_fellowship_ambassador_core.rs
@@ -176,7 +176,7 @@ impl<T: frame_system::Config> pallet_core_fellowship_ambassador::WeightInfo for 
 	/// Storage: `AmbassadorCollective::IdToIndex` (r:0 w:6)
 	/// Proof: `AmbassadorCollective::IdToIndex` (`max_values`: None, `max_size`: Some(54), added: 2529, mode: `MaxEncodedLen`)
 	/// The range of component `r` is `[1, 6]`.
-	fn promote_fast(r: u32, ) -> Weight {
+	fn promote_fast(r: u16, ) -> Weight {
 		// Proof Size summary in bytes:
 		//  Measured:  `65968`
 		//  Estimated: `69046 + r * (2489 ±0)`
